+++ CHANGELOG.md
@@ -110,6 +110,8 @@ Adding a new version? You'll need three changes:
   [#3121](https://github.com/Kong/kubernetes-ingress-controller/pull/3121)
 - Routes support annotations for path handling.
   [#3121](https://github.com/Kong/kubernetes-ingress-controller/pull/3121)
+- Warning events are recorded when CA secrets cannot be properly translated into Kong configuration.  
+  [#3125](https://github.com/Kong/kubernetes-ingress-controller/pull/3125)
 
 ### Fixed